9 ДЗ (#23)
9 ДЗ (#23)

+++ app/src/main/java/ru/kavunov/mtsproject/bd/BdModel.kt
@@ -35,10 +35,13 @@ data class MovieTable(
     val movId: Long,
     val title: String,
     val description: String,
-    val rateScore: Int,
+
+    val rateScore: Float,
     val ageRestriction: String,
     val imageUrl: String,
-)
+
+
+    )
 
 @Entity(tableName = "ActorModel")
 data class ActorTable(
